# docs/TalkRoomsOpenApi.yaml
openapi: 3.0.1
info:
  title: TalkRooms API
  description: 'System to manage message channels and send messages'
  termsOfService: https://swagger.io/terms/
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
...
paths:
  /user/register/{invitationId}:
    post:
      tags:
        - Users
      summary: Register a new user with invitation
      parameters:
        - name: invitationId
          in: path
          description: Invitation identifier
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        description: Register a new user given a username and password
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterInput'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: A user with such username already exists
        '500':
          description: Internal server error
  /user/register:
    post:
      tags:
        - Users
      summary: Register a new user
      requestBody:
        description: Register a new user given a username and password
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterInput'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: A user with such username already exists
        '500':
          description: Internal server error
  /user/login:
    post:
      tags:
        - Users
      summary: Retrieve a id, a token and a username given a username and a password, a login operation
      requestBody:
        description: Username and password of user
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticatedUser'
      responses:
        '201':
          description: User details
        '400':
          description: Username missing or invalid
        '404':
          description: A user with such username does not exists
        '500':
          description: Internal server error
  /user/logout:
    post:
      tags:
        - Users
      security:
        - bearerAuth: [ ]
      summary: Logout a user
      responses:
        '200':
          description: User details
        '401':
          description: Unauthorized
        '404':
          description: A user with such username does not exists
        '500':
          description: Internal server error
  /user/edit/username:
    put:
      tags:
        - Users
      security:
        - bearerAuth: [ ]
      summary: Change username
      requestBody:
        description: New username
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsernameUpdateInput'
      responses:
        '200':
          description: Username changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: A user with such username already exists
        '500':
          description: Internal server error
  /user/{id}:
    get:
      security:
        - bearerAuth: []
      tags:
        - Users
      summary: Get details of a user
      description: Returns a single user
      parameters:
        - name: id
          in: path
          description: User identifier
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Username missing or invalid
        '401':
          description: Unauthorized

        '404':
          description: User not found
        '500':
          description: Internal server error
  /user/search/{username}:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Users
      summary: Search user by username
      description: Returns a list of players by name
      parameters:
        - name: username
          in: path
          description: username
          required: true
          schema:
            type: string
            example: Bob123
        - name: limit
          in: query
          description: number of users to return
          required: false
          schema:
            type: integer
            example: 20
        - name: skip
          in: query
          description: number of users to skip
          required: false
          schema:
            type: integer
            example: 0
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
        '400':
          description: Invalid username
        '401':
          description: Unauthorized
        '404':
          description: User not found
        '500':
          description: Internal server error
...
  /channels/user/{userId}:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Channels
      summary: Get channels by user id
      description: Returns a list of channels
      parameters:
        - name: userId
          in: path
          description: user identifier
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Channels found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelList'
        '400':
          description: Invalid user
        '401':
          description: Unauthorized
        '404':
          description: User not found
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserIdentifiers:
      type: object
      description: User details
      required:
        - id
        - username
      properties:
        id:
          type: integer
          description: user id
          example: 1
        username:
          type: string
          description: username
          example: Bob123
...
    ChannelOutputModel:
      type: object
      description: Data that represents a channel
      required:
        - id
        - name
        - creator
        - visibility
      properties:
        id:
          type: integer
          description: channel identifier
          example: 1
        creator:
          $ref: '#/components/schemas/UserIdentifiers'
        name:
          type: string
          description: channel name
          example: DAW
        visibility:
          type: string
          description: channel visibility
          enum: [PUBLIC, PRIVATE]
          example: PUBLIC
    ChannelList:
      type: object
      properties:
        nChannels:
          type: integer
          description: number of results
          example: 20
        channels:
          type: array
          items:
            $ref: '#/components/schemas/ChannelOutputModel'
...
    UserList:
      type: object
      properties:
        nUsers:
          type: integer
          description: number of results
          example: 20
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserIdentifiers'
...
    UsernameUpdateInput:
      type: object
      description: User new username
      required:
        - newUsername
      properties:
        newUsername:
          type: string
          description: new username
          example: Bob1234
    UserRegisterInput:
      type: object
      description: User details for register
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          description: User username
          example: Bob
        email:
          type: string
          description: User email
          example: bob@example.com
        password:
          type: string
          description: User password
          example: Strong_but_known_password1234
...
    User:
      type: object
      description: User details
      required:
        - id
        - username
        - email
      properties:
        id:
          type: integer
          description: user identifier
        username:
          type: string
          description: user username
          example: Bob
        email:
          type: string
          description: user email
          example: bob@example.pt
    AuthenticatedUser:
      type: object
      description: User details
      required:
        - user
        - token
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
          description: session token
          example: absbadbdaksjdad_aksdjakld1223
...